+++ src/app.module.ts
@@ -6,6 +6,7 @@ import { AppService } from './app.service.js'
 import { AuthModule } from './core/auth/auth.module.js'
 import { ConfigModule } from './core/config/config.module.js'
 import { LoggerModule } from './core/logger/logger.module.js'
+import { SchedulerModule } from './core/scheduler/scheduler.module.js'
 import { AccessoriesModule } from './modules/accessories/accessories.module.js'
 import { BackupModule } from './modules/backup/backup.module.js'
 import { ChildBridgesModule } from './modules/child-bridges/child-bridges.module.js'
@@ -23,6 +24,7 @@ import { UsersModule } from './modules/users/users.module.js'
   imports: [
     ConfigModule,
     LoggerModule,
+    SchedulerModule,
     AuthModule,
     PluginsModule,
     ServerModule,