+++ src/core/config/config.service.ts
@@ -215,9 +215,10 @@ export class ConfigService {
         },
         plugins: {
           hideUpdatesFor: this.ui.plugins?.hideUpdatesFor || [],
-          hidePairingAlerts: this.ui.plugins?.hidePairingAlerts || [],
           alwaysShowBetas: Boolean(this.ui.plugins?.alwaysShowBetas),
         },
+        scheduledRestartCron: this.ui.scheduledRestartCron || null,
+        bridges: this.ui.bridges || [],
         linux: {
           shutdown: this.ui.linux?.shutdown,
           restart: this.ui.linux?.restart,