+++ ui/src/app/core/settings.interfaces.ts
@@ -2,6 +2,12 @@ export interface FeatureFlags {
   [key: string]: boolean
 }
 
+export interface BridgeConfig {
+  username: string
+  hideHapAlert?: boolean
+  scheduledRestartCron?: string
+}
+
 export interface EnvInterface {
   platform: 'darwin' | 'win32' | 'linux' | 'freebsd'
   enableAccessories: boolean
@@ -48,9 +54,10 @@ export interface EnvInterface {
   }
   plugins?: {
     hideUpdatesFor?: string[]
-    hidePairingAlerts?: string[]
     alwaysShowBetas?: boolean
   }
+  scheduledRestartCron?: string | null
+  bridges?: BridgeConfig[]
   linux?: {
     shutdown?: string
     restart?: string